refactor: update import paths for IRol interface
