+++ src/pages/privileges/outlets/roles/interface.tsx
@@ -16,7 +16,7 @@ import { Menu } from "@components/navigation/Menu";
 import { privilegeOptionsConfig } from "@pages/privileges/outlets//options/config/privileges.config";
 import { LoadingApp } from "@pages/login/outlets/LoadingApp";
 
-import { IRol } from "./add-role/forms/types/types";
+import { IRol } from "./types";
 import { menuInvitationLinks } from "./config/MenuAddRole";
 import {
   RolesBreakPointsConfig,
