+++ src/pages/privileges/outlets/roles/types.ts
